# src/Bff/Example.Bff/Controllers/Knowledge/KnowledgeMaterialsController.cs
using SuperApp.Framework.Infrastructure.Api;
using Example.Bff.Clients.Knowledge;
using Microsoft.AspNetCore.Mvc;
using System.Net.Mime;

namespace Example.Bff.Controllers.Knowledge;

/// <summary>
/// Public API of the Example experience for <c>Materials</c> of the Knowledge service: the module calls these endpoints, and the
/// BFF passes each call on to the Knowledge service with the user's token (ADR-0038, ADR-0039, ADR-0040).
/// </summary>
/// <remarks>
/// Pass-through endpoints: request and response bodies are the service's contract types generated by Refitter, and the service's
/// answer, including its error status and <c>code</c>, is relayed unchanged (<c>DownstreamResponseExtensions</c>). Authorization
/// (scopes, resource rules) is decided by the Knowledge service. When the service cannot be reached the BFF answers
/// <c>503 downstream.unavailable</c> or <c>504 downstream.timeout</c>.
/// </remarks>
/// <param name="knowledge">Client of the Knowledge service, generated from its committed OpenAPI contract.</param>
[ApiController]
[Tags("Knowledge – Materials")]
public sealed class KnowledgeMaterialsController(IKnowledgeApi knowledge) : ControllerBase
{
    /// <summary>Returns a page of published materials, newest first, optionally filtered by category and type.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.read</c>. Only materials in status <c>Published</c> are returned, ordered by publication
    /// date descending. The summaries do not contain the content; use <c>GET /v1/materials/{id}</c> for it.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="categoryId">Optional filter: only materials assigned to this category. An unknown category yields an empty page.</param>
    /// <param name="type">Optional filter by material type: <c>Article</c>, <c>Video</c> or <c>Podcast</c>.</param>
    /// <param name="page">Page number starting at 1; values lower than 1 are treated as 1.</param>
    /// <param name="pageSize">Page size, 20 by default, clamped to the range 1–100.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="200">The requested page (possibly empty).</response>
    /// <response code="400">An invalid filter value, e.g. an unknown material type.</response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.read</c> (<c>auth.missing_scope</c>).</response>
    [HttpGet("v1/knowledge/materials")]
    [ProducesResponseType<PagedResultOfMaterialSummaryDto>(StatusCodes.Status200OK, MediaTypeNames.Application.Json)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> List([FromQuery] System.Guid? categoryId, [FromQuery] MaterialType? type, [FromQuery] int? page, [FromQuery] int? pageSize, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsListAsync(categoryId, type, page, pageSize, cancellationToken));

    /// <summary>Creates a new material in status <c>Draft</c>, without content and categories.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.write</c> (editor). The type cannot be changed later. Title: required, at most 200 characters;
    /// description: optional, at most 2000 characters (both trimmed). The main media (an absolute <c>https</c> URL of the video or audio
    /// file and its non-negative duration in seconds) applies only to <c>Video</c> and <c>Podcast</c>; an article cannot have it.
    /// A duration without a URL is rejected. Videos and podcasts may be created without main media, but need it to be published.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="body">Type, title, description and optional main media of the material.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="201">The material was created; the body contains its identifier.</response>
    /// <response code="400">
    /// Validation error: invalid type, title, description, URL or duration (<c>validation.failed</c>,
    /// <c>knowledge.material.invalid_title</c>, <c>knowledge.material.invalid_description</c>, <c>knowledge.url.invalid</c>,
    /// <c>knowledge.material.media_not_allowed</c>, <c>knowledge.material.invalid_duration</c>).
    /// </response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    [HttpPost("v1/knowledge/materials")]
    [ProducesResponseType<CreatedResponse>(StatusCodes.Status201Created, MediaTypeNames.Application.Json)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> Create([FromBody] CreateMaterial body, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsCreateAsync(body, cancellationToken));

    /// <summary>Returns the details of a material with its complete block content.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.read</c>. A reader sees only published materials; an editor
    /// (scope <c>knowledge.catalog.write</c>) also sees drafts and archived materials. Published materials are served from a cache that is
    /// invalidated whenever the material changes, so readers see changes immediately after a successful write.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="200">The material details.</response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.read</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">
    /// The material does not exist or is not visible to the caller, e.g. a draft requested by a reader (<c>knowledge.material.not_found</c>).
    /// </response>
    [HttpGet("v1/knowledge/materials/{materialId:guid}")]
    [ProducesResponseType<MaterialDetailsDto>(StatusCodes.Status200OK, MediaTypeNames.Application.Json)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> Get(System.Guid materialId, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsGetAsync(materialId, cancellationToken));

    /// <summary>Changes the title, description and main media of a material.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.write</c> (editor). The same rules as on creation; all fields are replaced, and <c>null</c>
    /// removes the description or the main media. A published video or podcast cannot lose its main media. An archived material cannot
    /// be changed. Changes are not announced to other services.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <param name="body">The new details of the material.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="204">The details were changed.</response>
    /// <response code="400">
    /// Validation error: invalid title, description, URL, duration or an empty identifier (<c>validation.failed</c>,
    /// <c>knowledge.material.invalid_title</c>, <c>knowledge.material.invalid_description</c>, <c>knowledge.url.invalid</c>,
    /// <c>knowledge.material.media_not_allowed</c>, <c>knowledge.material.invalid_duration</c>).
    /// </response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">The material does not exist (<c>knowledge.material.not_found</c>).</response>
    /// <response code="409">
    /// The material was changed by another request between reading and saving (<c>persistence.concurrency_conflict</c>);
    /// reload it and retry.
    /// </response>
    /// <response code="422">
    /// The material is archived (<c>knowledge.material.archived</c>), or the main media would be removed from a published video or
    /// podcast (<c>knowledge.material.main_media_required</c>).
    /// </response>
    [HttpPut("v1/knowledge/materials/{materialId:guid}")]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status409Conflict, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status422UnprocessableEntity, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> UpdateDetails(System.Guid materialId, [FromBody] UpdateMaterialDetailsRequest body, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsUpdateDetailsAsync(materialId, body, cancellationToken));

    /// <summary>Replaces the whole block content of a material.</summary>
    /// <remarks>
    /// Required scope: knowledge.catalog.write (editor). The content is a tree of blocks distinguished by the type field.
    /// The main rules: at most 500 blocks in total (children included); each block type allows only specific child types; lists can be
    /// nested at most 3 levels deep and toggles 2 levels; every URL must be an absolute https URL (links in text may also use
    /// mailto:); Embed blocks are accepted only from allowed providers; Image blocks require an alternative text.
    ///     The reading time is recalculated from the new content. A published material cannot be emptied; an archived material cannot be
    /// changed.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <param name="body">The new content: top-level blocks with their child blocks.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="204">The content was replaced.</response>
    /// <response code="400">
    /// The content breaks a block rule (<c>knowledge.content.invalid_block</c>; the message names the path of the offending block, e.g.
    /// <c>blocks[2].children[0]</c>), the block list is missing or the identifier is empty (<c>validation.failed</c>), or the body is not
    /// valid content JSON (e.g. an unknown block <c>type</c>).
    /// </response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">The material does not exist (<c>knowledge.material.not_found</c>).</response>
    /// <response code="409">
    /// The material was changed by another request between reading and saving (<c>persistence.concurrency_conflict</c>);
    /// reload it and retry.
    /// </response>
    /// <response code="422">
    /// The material is archived (<c>knowledge.material.archived</c>), or it is published and the new content is empty
    /// (<c>knowledge.material.content_required</c>).
    /// </response>
    [HttpPut("v1/knowledge/materials/{materialId:guid}/content")]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status409Conflict, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status422UnprocessableEntity, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> ReplaceContent(System.Guid materialId, [FromBody] ReplaceContentRequest body, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsReplaceContentAsync(materialId, body, cancellationToken));

    /// <summary>Replaces the set of categories of a material.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.write</c> (editor). At most 20 distinct categories, all of which must exist; duplicates in the
    /// request are ignored and an empty list removes all assignments. An archived material cannot be changed.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <param name="body">Identifiers of the categories.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="204">The categories of the material were set.</response>
    /// <response code="400">
    /// Validation error: missing list or empty identifiers (<c>validation.failed</c>), unknown categories (<c>knowledge.category.unknown</c>)
    /// or too many categories (<c>knowledge.material.too_many_categories</c>).
    /// </response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">The material does not exist (<c>knowledge.material.not_found</c>).</response>
    /// <response code="409">
    /// The material was changed by another request between reading and saving (<c>persistence.concurrency_conflict</c>);
    /// reload it and retry.
    /// </response>
    /// <response code="422">The material is archived (<c>knowledge.material.archived</c>).</response>
    [HttpPut("v1/knowledge/materials/{materialId:guid}/categories")]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status409Conflict, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status422UnprocessableEntity, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> SetCategories(System.Guid materialId, [FromBody] SetCategoriesRequest body, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsSetCategoriesAsync(materialId, body, cancellationToken));

    /// <summary>Publishes a material, making it visible to readers.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.write</c> (editor). Publishing requires content, and a video or podcast also requires main
    /// media. The first publication records the publication date and publishes the integration event <c>MaterialPublishedV1</c>.
    /// The operation is idempotent: publishing an already published material succeeds without changes and without a new event.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="204">The material is published.</response>
    /// <response code="400">Empty identifier (<c>validation.failed</c>).</response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">The material does not exist (<c>knowledge.material.not_found</c>).</response>
    /// <response code="409">
    /// The material was changed by another request between reading and saving (<c>persistence.concurrency_conflict</c>);
    /// reload it and retry.
    /// </response>
    /// <response code="422">
    /// The material is archived, has no content or lacks the main media
    /// (<c>knowledge.material.archived</c>, <c>knowledge.material.content_required</c>, <c>knowledge.material.main_media_required</c>).
    /// </response>
    [HttpPost("v1/knowledge/materials/{materialId:guid}/publish")]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status409Conflict, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status422UnprocessableEntity, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> Publish(System.Guid materialId, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsPublishAsync(materialId, cancellationToken));

    /// <summary>Archives a material: it is hidden from readers and removed from users' favorites.</summary>
    /// <remarks>
    /// Required scope: <c>knowledge.catalog.write</c> (editor). Archiving is irreversible (an archived material can be neither changed nor
    /// published again) and publishes the integration event <c>MaterialArchivedV1</c>. Favorites pointing to the material are removed
    /// asynchronously, shortly after the response; until then they are already hidden from the favorites list. Completion marks are kept.
    /// The material stays in the collections that contain it but is hidden from readers there. The operation is idempotent: archiving
    /// an already archived material succeeds without changes.
    /// </remarks>
    /// <param name="cancellationToken">Cancellation of the HTTP request.</param>
    /// <param name="materialId">Identifier of the material.</param>
    /// <returns>The service's answer, relayed unchanged.</returns>
    /// <response code="204">The material is archived.</response>
    /// <response code="400">Empty identifier (<c>validation.failed</c>).</response>
    /// <response code="401">Missing, expired or invalid access token.</response>
    /// <response code="403">The token lacks the scope <c>knowledge.catalog.write</c> (<c>auth.missing_scope</c>).</response>
    /// <response code="404">The material does not exist (<c>knowledge.material.not_found</c>).</response>
    /// <response code="409">
    /// The material was changed by another request between reading and saving (<c>persistence.concurrency_conflict</c>);
    /// reload it and retry.
    /// </response>
    [HttpPost("v1/knowledge/materials/{materialId:guid}/archive")]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ValidationProblemDetails>(StatusCodes.Status400BadRequest, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status403Forbidden, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status404NotFound, MediaTypeNames.Application.ProblemJson)]
    [ProducesResponseType<Microsoft.AspNetCore.Mvc.ProblemDetails>(StatusCodes.Status409Conflict, MediaTypeNames.Application.ProblemJson)]
    public async Task<IActionResult> Archive(System.Guid materialId, CancellationToken cancellationToken) =>
        this.ToActionResult(await knowledge.MaterialsArchiveAsync(materialId, cancellationToken));
}