# src/Bff/Example.Bff/openapi/Example.Bff_public.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Example.Bff | public",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/knowledge/materials": {
      "get": {
        "tags": [
          "Knowledge – Materials"
        ],
        "summary": "Returns a page of published materials, newest first, optionally filtered by category and type.",
        "description": "Required scope: `knowledge.catalog.read`. Only materials in status `Published` are returned, ordered by publication\r\ndate descending. The summaries do not contain the content; use `GET /v1/materials/{id}` for it.",
        "operationId": "KnowledgeMaterials_List",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "description": "Optional filter: only materials assigned to this category. An unknown category yields an empty page.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Optional filter by material type: `Article`, `Video` or `Podcast`.",
            "schema": {
              "$ref": "#/components/schemas/MaterialType"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number starting at 1; values lower than 1 are treated as 1.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Page size, 20 by default, clamped to the range 1–100.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested page (possibly empty).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgePagedResultOfMaterialSummaryDto"
                }
              }
            }
          },
          "400": {
            "description": "An invalid filter value, e.g. an unknown material type.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `knowledge.catalog.read` (`auth.missing_scope`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Knowledge – Materials"
        ],
        "summary": "Creates a new material in status `Draft`, without content and categories.",
        "description": "Required scope: `knowledge.catalog.write` (editor). The type cannot be changed later. Title: required, at most 200 characters;\r\ndescription: optional, at most 2000 characters (both trimmed). The main media (an absolute `https` URL of the video or audio\r\nfile and its non-negative duration in seconds) applies only to `Video` and `Podcast`; an article cannot have it.\r\nA duration without a URL is rejected. Videos and podcasts may be created without main media, but need it to be published.",
        "operationId": "KnowledgeMaterials_Create",
        "requestBody": {
          "description": "Type, title, description and optional main media of the material.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeCreateMaterial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeCreateMaterial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeCreateMaterial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The material was created; the body contains its identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error: invalid type, title, description, URL or duration (`validation.failed`,\r\n`knowledge.material.invalid_title`, `knowledge.material.invalid_description`, `knowledge.url.invalid`,\r\n`knowledge.material.media_not_allowed`, `knowledge.material.invalid_duration`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `knowledge.catalog.write` (`auth.missing_scope`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/v1/knowledge/materials/{materialId}/publish": {
      "post": {
        "tags": [
          "Knowledge – Materials"
        ],
        "summary": "Publishes a material, making it visible to readers.",
        "description": "Required scope: `knowledge.catalog.write` (editor). Publishing requires content, and a video or podcast also requires main\r\nmedia. The first publication records the publication date and publishes the integration event `MaterialPublishedV1`.\r\nThe operation is idempotent: publishing an already published material succeeds without changes and without a new event.",
        "operationId": "KnowledgeMaterials_Publish",
        "parameters": [
          {
            "name": "materialId",
            "in": "path",
            "description": "Identifier of the material.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The material is published."
          },
          "400": {
            "description": "Empty identifier (`validation.failed`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid access token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The token lacks the scope `knowledge.catalog.write` (`auth.missing_scope`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The material does not exist (`knowledge.material.not_found`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The material was changed by another request between reading and saving (`persistence.concurrency_conflict`);\r\nreload it and retry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The material is archived, has no content or lacks the main media\r\n(`knowledge.material.archived`, `knowledge.material.content_required`, `knowledge.material.main_media_required`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KnowledgePagedResultOfMaterialSummaryDto": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KnowledgeMaterialSummaryDto"
            },
            "description": "Items of the requested page, in the query's order; empty when the page is beyond the last item."
          },
          "page": {
            "type": "integer",
            "description": "Number of the returned page, starting at 1.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Effective maximum number of items per page after clamping (1 to int Paging.MaxPageSize); the last page may contain fewer.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Number of items matching the query across all pages; clients compute the page count as `ceil(TotalCount / PageSize)`.",
            "format": "int32"
          }
        },
        "description": "One page of a list query, together with the information a client needs to render paging (total count, page number, page size)."
      },
...
      "MaterialType": {
        "enum": [
          "Article",
          "Video",
          "Podcast",
          null
        ]
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code, e.g. knowledge.material.archived; clients branch on it (ADR-0015)."
          },
          "traceId": {
            "type": "string",
            "description": "Trace identifier of the request, used to find its traces and logs."
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT access token issued by the CIAM, passed unchanged by the edge gateway and BFF.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}